Run a full garbage colletion after a test step.
This turned out to be necessary to get rid of any existing objects which
block a re-run if they still exist. The problematic situation was a test case
which used a plugin which opened a USB connection. The test case stopped after
an error was detected. So the close method of the USB plugin was not called.
Although the plugin only existed in the "run" method of the test and the
plugin destructor closes the USB connection, the plugin object still existed
after the "run" method.