Split system and test execution parameter.
Now the test execution parameter are set to the response of the
"select_serial..." dialog. The system parameter are read from the
"tests.xml" and extended with the current serial number under test
as "serial".
The old implementation had only the system parameter, which was a
mixture of the "select_serial..." response and the current serial
number under test.