Use one field for device number, revision and serial.
The first serial number and the amount of boards to test had
separate input fields. Now there is one field which combines the
device number, hardware revision and serial number. The device
number has 7 digits, the hardware revision is one character of
either 0-9 or a-z and the serial number has 5 or 6 digits. This is
the same as a standard Hilscher matrix label. For example the
device number 1234567, hardware revision 8 and serial number 23456
result in the value 1234567823456 .