Check the contents of the matrix label with a function.
The test description was exteded by a configuration block. It provides the
3 parameter LabelValidationFunction, LabelValidationDataTyp and LabelValidationData.
The LabelValidationFunction is either the body of a Javascript function
which executes the test of a matrix label. It gets 2 arguments: first is
the label contents as a string, second is user data necessary to do the test.
LabelValidationDataTyp is the type of the user data. It can be JSON,
CSV and STRING. It defines how the LabelValidationData parameter is
converted. JSON tries to parse the parameter as JSON and passes the
resulting object as the 2nd parameter to the label validation function.
CSV separates the input by comma and removes leading and trailing whitespace
from every piece. The resulting array is passed as the 2nd parameter to the
label validation function. The last option STRING does no conversion.
LabelValidationFunction accepts the 2 magic keywords to select built-in
functions. "NONE" does no label validation. With "PREFIX-LIST" the
contents of the matrix label must start with a prefix defined either as CSV
prefix1, prefix2, prefix3,... or JSON ['prefix1','prefix2','prefix3'].
The default settings for the configuration parameters deactivate the label
validation with the "NONE" keyword in the LabelValidationFunction parameter.

Example for the test.xml:

        <Configuration>
                <Parameter name="LabelValidationFunction">PREFIX-LIST</Parameter>
                <Parameter name="LabelValidationDataTyp">CSV</Parameter>
                <Parameter name="LabelValidationData">28014003, 28014004, 28015003, 28015103</Parameter>
        </Configuration>

This activates label validation and accepts only labels startting with the
strings "28014003", "28014004", "28015003" or "28015103". All other labels are
rejected as invalid.

Please note that you can disable label validation and production number
validation in the advanced settings.