Specify additional input fields in the serial number dialog.
The new parameter "AdditionalInputs" in the configuration section of
the tests.xml defines the inputs. Example:

<?xml version="1.0" encoding="UTF-8"?>
<MuhkuhTest title="Board test" subtitle="a great board">
        <Configuration>
                <Parameter name="AdditionalInputs">[{"id": "aoi_top", "label": "AOI Label Top side", "required": "true", "regexp": "^[0-9]+$"}]</Parameter>
        </Configuration>
...

The example creates a new input field with the label "AOI Label Top side"
which is required and must match the regular expression "^[0-9]+$", which
is one or more numbers. The result is stored in the system parameter
"aoi_top", which can be accessed as a system parameter like this:

  <Connection name="aoi_top">system:aoi_top</Connection>